refactor: reorganize developer CI-subpackage, isolate gen-op-dataset into datasets/pureop.py
