+++ developer/__init__.py
@@ -1,0 +1,1 @@
+# we need it to generate documentation from helper scripts
